requirements.txt
fastapi
uvicorn
celery
redis
sqlalchemy
psycopg2-binary
transformers
torch
python-dotenv
pygithub
pydantic
python-jose
passlib
pytest
httpx